+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * Support callable for "money" model table modification.
 * Added multi server support with end-to-end authorization.
 * Added command for generate admin key.
+* Added total row for model table.
 
 ## [6.4.1](https://github.com/bfg-s/admin/compare/6.4.0...6.4.1) - 2024-11-06
 * New dev theme for the admin panel.